+++ backend/canisters/local_user_index/impl/src/updates/c2c_user_canister.rs
@@ -7,11 +7,8 @@ use local_user_index_canister::UserEvent;
 use local_user_index_canister::c2c_user_canister::*;
 use stable_memory_map::StableMemoryMap;
 use std::cell::LazyCell;
-use types::{
-    BotEvent, BotInstallationLocation, BotLifecycleEvent, Notification, StreakInsuranceClaim, StreakInsurancePayment,
-    TimestampMillis, UserId,
-};
-use user_index_canister::BotInstallationUpdated;
+use types::{BotEvent, BotLifecycleEvent, Notification, StreakInsuranceClaim, StreakInsurancePayment, TimestampMillis, UserId};
+use user_index_canister::{BotInstalled, BotUninstalled};
 
 #[update(guard = "caller_is_local_user_canister", msgpack = true)]
 #[trace]
@@ -94,9 +91,10 @@ fn handle_event<F: FnOnce() -> TimestampMillis>(
                 && let BotEvent::Lifecycle(BotLifecycleEvent::Installed(event)) = &bot_notification.event
             {
                 state.push_event_to_user_index(
-                    crate::UserIndexEvent::BotInstallationUpdated(Box::new(BotInstallationUpdated {
+                    crate::UserIndexEvent::BotInstalled(Box::new(BotInstalled {
                         bot_id: bot_notification.recipients[0],
-                        location: BotInstallationLocation::User(user_id.into()),
+                        location: event.location,
+                        installed_by: event.installed_by,
                         granted_permissions: event.granted_command_permissions.clone(),
                         granted_autonomous_permissions: event.granted_autonomous_permissions.clone(),
                     })),