+++ backend/canisters/user_index/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Allow platform operators to block usernames via regex patterns ([#8719](https://github.com/open-chat-labs/open-chat/pull/8719))
+- Add endpoint for bots to query their installation locations ([#8731](https://github.com/open-chat-labs/open-chat/pull/8731))
 
 ## [[2.0.1933](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1933-user_index)] - 2025-12-05